Adding alerts app
Instantiating a new alerts app and moving the alert functionality from
events/alert.py to alerts/old_alert.py. This commit forms a starting
point where the old alert functionality is still in place and we can
rework the alert infrastructure to be more modularized, flexible,
and adapted to superevents.