Move some imports in superevents/serializers.py
Some imports in superevents/serializers.py were moved to
function-level.  This is because several functions from
superevents/utils.py are used in the serializers for
deserialization (object creation), but the serializers are also
used in the alerts app to serialize objects to dicts.  This created
a circular import issue.

We also fix some of the reverses so that they work without a
request object being present.