Restrict tag name characters
If a tag's name has some "uncommon" characters in it, like
parentheses (or probably others), it will cause errors for the
javascript which renders the event pages, since it tries to use
the tag name in the element id for the delete button.  We now require
that tags only use 0-9a-zA-Z_\- in their names. May have to go through
and delete some tags which use them.