Middleware for checking client version
Adding middleware which checks the client version (for everything
under /api*) by using the User-Agent header. Obviously requires the
client to set the header appropriately as well.  Allows access
when the header is set but 'gracedb-client' is not found in it
since browsers have to access the web browsable API.