Updating superevent date handling with 2 digit yr
Fixing ambiguity in superevent date IDs due to use of a 2 digit year.
datetime uses some range like 1969-2068, but we would prefer something
more GPS-centric, so we force it to be 1980-2079.

This will need to be dealt with in advance of 2080, probably
will need to convert to 4 digit years.  Shouldn't be too hard,
will likely need to just make the switch, but maintain legacy
compatibility with legacy URLs, search features, and a check in
Superevent.get_filter_kwargs_for_date_id_lookup, which
points to superevents between 1980-2079 if a 2 digit year is given.