Moving events and superevents APIs into api app
Moved events.api and superevents.api into the new api app.
We want to have a versioned API, so these are now in
api.v1.events and api.v1.superevents, respectively. Part of this
change is that we require a custom reverse function (api_reverse)
for handling versioning. Much of the required code changes for this
migration include converting reverses to use this function and
changing relative imports appropriately.