Bugfix for superevent VOEvents
Since we use the preferred event to construct much of the VOEvent
for a superevent, we have to select the event's subclass (note that this
is done for events by the event_and_auth_required decorator, which calls
Event.getByGraceid).  This was preventing certain subclass-specific
attributes from being added to the VOEvents, like ProbHasNS,
ProbHasRemnant, etc.