Abstracting/generalizing several classes
Many of the classes (permissions, mixins, etc.) in the superevents
API follow a general pattern of
    1. Abstract customized version of a DRF class
    2. Class-specific, customized version of #1
    3. (optional) Implemented, class-specific version of #2
We want all classes in #1 to either be in the root api directory
or the root versioned (api.v*) directory.  #2 and #3 classes should
be in api.v*.superevents. This commit just moves some of the classes
in #1 and #2 around so that they can eventually be used in the events
API.