Adding handling for nan values in eff_distance
gstlal wants to set eff_distance to nan when they don't have a
meaningful value for it.  Unfortunately, we can't save that in the
database because MySQL numeric types can't deal with nans. So we
check for nans in this field and set it to NULL if it is nan,
before saving it in the database.