Update superevent file list and detail web views
Updated the web view for the superevent file list to be a
class-based view which uses view permission-based filtering. Also
updated the file links on the list page to point to the file download
page on the API so as to reduce redundancy.  Removed the old file
list and detail web views.