Show log and emo issuer full name in web view
The serializers for logs and emobservations (for both events and
superevents) now give the user's full name for web views (unless
it's blank; then fallback to username) and the username otherwise.
We check for web view 'status' by seeing whether it is an AJAX
request. This matches the previous behavior before the development
of the superevents features.