Make subject of X509Cert longer
Some X509 certificate subjects which were obtained from the LIGO
LDAP were being truncated upon insertion into the database since
they were longer than the upper limit imposed here of 200 characters.
I asked the LIGO auth team for guidance as to whether there is any
kind of upper limit imposed at that level, but did not receive any
feedback.  So I am increasing it to 300 characters here and hoping
that is enough that we will never have to revisit this issue.  It will
also be necessary to clean up any certificates in the database which
are truncated.