Change how tags are created in unit tests
Have to use 'get_or_create()' for Tag creation in unit tests.
Previously any tags that were created by data migrations were
always wiped after the initial migrations ran to set up the test
database due to the fact that we use MyISAM.  Now initial data
seems to be cleared only after the first test class runs.  So if
you run the full set of unit tests, it works fine, but if you only
run a class which includes creating a "real" Tag which is part
of the data migrations, it fails since the Tag is already in the
database.

I don't know when this changed - I tried using a previous tag of
the code (2.2.0) and the last version of Django that we used
(1.11.16), but no luck.  Really the solution should be to go away
from MyISAM or even MySQL altogether, but that will have to come
at a later date.