Bugfix event log uploads without file
Have seen issues with "column 'filename' cannot be null" warnings
in the logs recently.  Found the issue in the events API, but
it's not obvious why this started now.  Possibly the upgrade to
Django 1.11.18?  Could have been a MariaDB upgrade too, but haven't
checked the logs to see if that happened.