Increase gunicorn keep-alive timeout for AWS deployment
Suggested keepalive time is 1-5 seconds for standard deployment,
but it should be much higher for deployments behind a load
balancer. We are hoping this might resolve the SSLEOFErrors which
have been seen on the AWS deployment.  Traefik's keepalive time
is hard-coded to 90 seconds so we want gunicorn's to be longer than
that, so we set it to 100 s.