Populate database columns for existing VOEvents
Data migration to populate new database columns for existing
VOEvents. We have to read the files in and try to extract the
parameter data, handling all of the changes to the file format
which have occurred over time.

I did a lot of testing by copying over the production database
and associated VOEvent files for all non-Test, non-MDC events
and superevents to make sure we are accounting for all changes
in VOEvent (file) schema that have happened over time.