Python 3: migrations
Not sure why these were generated. The models were not modified,
but maybe it was just because we are now using Python 3, and string
handling has changed compared to Python 2.  I think all of the fields
that were touched have strings in them either as a default option
or as a set of choices.