Restructuring AuthGroup to add KAGRA
KAGRA members send a different ldap group name (ldap_name) but the
current group stsructure doesn't allow for that. This is a pretty
invasive restructuring that basically adds a many-to-one relationship
map different ldap names to the same group. I think the tests are still failing,
but they should be restructured.