Applying a retry "bandaid"
We ran into issues where there were [Errno 5] Input/Output
errors, but only on AWS, and it was completely infrequent and
not reproducible. Great. When I would go back in and run the
same command or look for the same files/folders manually on
EFS, it was always there. So I'm trying this on playground and
test: just retry the offending command five times until it works.
I looked in the traceback, and it was in the list_files or retrieve
each time. Let's see!!