Upgrade pipeline image and services
I was running into mysql errors when porting my changes that
worked locally into the gitlab ci pipeline. There's this article:

https://stackoverflow.com/questions/52730817/django-migrate-error-mysql-exceptions-programmingerror-1064-you-have-an-err

that shows that django 2.1+ supports MySQL 5.6 and higher, but the
ci pipeline was running 5.5. Bump this up and see if it works, and
switch to the igwn container instead of the depreciated ligo one.