tweak public alerts page to include ER15 candidates
context: users can search a gpstime range for events and superevents
by searching by the observation/engineering period, which are hardcoded
in. the public page was changed last week to separate candidates by
observation period, as well as providing a direct anchor link to each
section. ER15 wasn't included at first (just O4/O3) because what are the
changes of a significant public candidate in the next week, right? well
here we are.