remove dependency on dojo javascript on superevent pages
the old detail script would wait for the dojo.parser to load,
which in turn waited for the main page to completely load before
enabling the advocate form. this was necessary when it was a dojo form,
but since switching to jquery+ajax, this is vestigial.

and there were corner cases where slow-loading images would prevent
advocates from clicking the ADVOK button without reloading. this
removes the limitation that the page be fully loaded, while still
requiring that the ajax is in place.